+++ src/frontend-models/base.js
@@ -782,7 +782,7 @@ export default class FrontendModelBase {
   /**
    * @template {typeof FrontendModelBase} T
    * @this {T}
-   * @param {string | string[]} sort - Sort definition(s), for example `"createdAt"` or `"-createdAt"`.
+   * @param {string | string[] | [string, string] | Array<[string, string]> | Record<string, any> | Array<Record<string, any>>} sort - Sort definition(s).
    * @returns {FrontendModelQuery<T>} - Query builder with sort definitions.
    */
   static sort(sort) {